Make both GNU as and NASM syntax versions
Just to get used to the two toolchains.
After doing this I think I've decided I want to work with NASM for the
bootloader. The intel syntax is much easier to type, and as long as the
project is mostly assembly I don't have to worry about toolchain
incompatibilities with gcc.