Looked into loading a ramdisk
It turns out that loading any old ramdisk and kernel is more difficult
than I imagined because the BIOS interrupt 0x15 (AH = 0x87) limitation
of 24 bit addresses can be reached.
This is the case of the random debian distribution I've been working
with.

At some point I'll have to try out compiling ext support in to the
kernel and keeping the partition sector the same, but for now I'll leave
it without having ever booted anything.